emacs-pager: Use sh for pager wrapper
Ruby is great and all, but not universally installed.  Rewrite the wrapper
in plain ol' shell to make it more portable.

This also changes the behavior of `emacs-pager` slightly in that it invokes
Emacs if it's not already running by using the `emacsclient` `-a` switch
with an empty string argument.